refactor: 유틸 함수 불러올 때 한 곳에 모아둔 곳으로부터 불러오게 해라
