+++ src/app/controller/ChargeCoin.js
@@ -1,7 +1,6 @@
 import ChargeCoinTap from '../view/ChargeCoinTap.js';
 import checkChargeAmount from '../asset/validation/checkChargeAmount.js';
-import getEnableChargeAmount from '../asset/util/getEnableChargeAmount.js';
-import distributeCoin from '../asset/util/distributeCoin.js';
+import { getEnableChargeAmount, distributeCoin } from '../asset/util/index.js';
 import { getChargeAmount, getCoins, setCoins } from '../localStorage/coin.js';
 
 export default class ChargeCoin {
